Use "Active" cases as current
Some countries (for instance Sweden and Belgium) does not report "recovered" cases, leading to "current" being "None".
Since we already have the field "Active" in the json file, it seems like a better choice to use that. 
But we are still falling back to the old method if we dont have a number for "Active" cases (although at least today it seems like that is present for all countries).